@@ -9,11 +9,13 @@
     <packaging>pom</packaging>
 
     <properties>
-        <java.version>17</java.version>
+        <java.version>21</java.version>
     </properties>
 
+    <!--
     <modules>
        <module>spring-boot-test-libraries-basics</module>
        <module>spring-boot-test-test-containers</module>
     </modules>
+    -->
 </project>
