@@ -5,7 +5,7 @@
     <parent>
         <groupId>org.springframework.boot</groupId>
         <artifactId>spring-boot-starter-parent</artifactId>
-        <version>3.0.5</version>
+        <version>3.5.3</version>
         <relativePath/> <!-- lookup parent from repository -->
     </parent>
     <groupId>io.github.jlmc</groupId>
@@ -24,15 +24,14 @@
     </developers>
 
     <properties>
-        <java.version>17</java.version>
+        <java.version>21</java.version>
         <testcontainers.version>1.17.6</testcontainers.version>
         <springdoc.version>1.6.15</springdoc.version>
         <mongock.version>5.2.4</mongock.version>
         <opencsv.version>5.7.1</opencsv.version>
-        <maven.jacoco.plugin.version>0.8.0</maven.jacoco.plugin.version>
+        <maven.jacoco.plugin.version>0.8.13</maven.jacoco.plugin.version>
         <maven.sonar.plugin.version>3.7.0.1746</maven.sonar.plugin.version>
         <org.mapstruct.version>1.5.4.Final</org.mapstruct.version>
-        <lombok.version>1.18.20</lombok.version>
         <lombok-mapstruct-binding.version>0.2.0</lombok-mapstruct-binding.version>
     </properties>
 
@@ -184,8 +183,8 @@
                 <artifactId>maven-compiler-plugin</artifactId>
                 <version>3.8.1</version>
                 <configuration>
-                    <source>17</source> <!-- depending on your project -->
-                    <target>17</target> <!-- depending on your project -->
+                    <source>21</source> <!-- depending on your project -->
+                    <target>21</target> <!-- depending on your project -->
                     <!-- See https://maven.apache.org/plugins/maven-compiler-plugin/compile-mojo.html -->
                     <!-- Classpath elements to supply as annotation processor path. If specified, the compiler   -->
                     <!-- will detect annotation processors only in those classpath elements. If omitted, the     -->
@@ -256,6 +255,7 @@
             <plugin>
                 <groupId>org.jacoco</groupId>
                 <artifactId>jacoco-maven-plugin</artifactId>
+                <version>${maven.jacoco.plugin.version}</version>
                 <configuration>
                     <excludes>
                         <exclude>io/github/jlmc/uploadcsv/migrations/**</exclude>
@@ -307,7 +307,7 @@
                 <plugin>
                     <groupId>org.jacoco</groupId>
                     <artifactId>jacoco-maven-plugin</artifactId>
-                    <version>0.8.7</version>
+                    <version>${maven.jacoco.plugin.version}</version>
                 </plugin>
                 <plugin>
                     <groupId>org.apache.maven.plugins</groupId>
